requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli == 1.27.165
botocore == 1.29.165 
flake8
python-dotenv>=0.5.1
numpy
pandas
matplotlib
seaborn
dvc
dvc-gs
dvc[gdrive]
google-cloud-storage
pydrive